Complete support for cgroups
This commit completes cgroups support for bfq-mq. In particular, it deals with
a sort of circular dependency introduced in blk-mq: the function
blkcg_activate_policy, invoked during scheduler initialization, triggers the
invocation of the has_work scheduler hook (before the init function is
finished). To adress this issue, this commit moves the invocation of
blkcg_activate_policy after the initialization of all the fields that could be
initialized before invoking blkcg_activate_policy itself. This enables has_work
to correctly return false, and thus to prevent the blk-mq stack from invoking
further scheduler hooks before the init function is finished.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>